Remove ReactInstrumentation.debugTool in production
We plan to add a separate profiling build in the future, so gating `ReactInstrumentation.debugTool` behind `__DEV__` conditions will soon be incorrect.

Here, we change the relevant conditions to check for `ReactInstrumentation.debugTool` instead. This is a runtime check, so this code will no longer be dead-code eliminated in production, at least not until we have a separate `__PROFILE__` gate. This will need further discussion and is not in the scope of this pull request, but at least all the callsites will be easy to find-replace now that they are separate from the `__DEV__` gates.

We also make a few other changes:

* `ReactInstrumentation.debugTool` is now `null` both in tests and in production. This leaves us confident that we don’t ship any devtools code in production, and that we can’t accidentally dispatch a devtool event without a gate check.
* Tests for *devtools themselves* now need to shim `process.env.NODE_ENV` to be `"development"` rather than `"test"` to enable the default debug tool.